+++ packages/react-aria-widgets/package.json
@@ -29,9 +29,9 @@
     "url": "git+https://github.com/charlie-yao/react-aria-widgets.git"
   },
   "scripts": {
-    "test-compile": "tsc -p tsconfig.test-compile.json",
+    "build:test": "tsc -p tsconfig.test-compile.json",
     "build": "webpack --config webpack.prod.js",
-    "build-dev": "webpack --config webpack.dev.js",
+    "build:dev": "webpack --config webpack.dev.js",
     "lint": "eslint --fix --ext .jsx,.js,.ts,.tsx ./",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
